추상화 처리
비즈니스 로직 메서드인 proc 을 추상메서드로 변경
